+++ src/components/form.ts
@@ -1,4 +1,4 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { Model } from 'modelsafe';
 
 import { FormDefinition } from '../definitions/form';
@@ -23,4 +23,20 @@ export class FormComponent {
    * In this case of creation, this could be an empty object.
    */
   @Input() data: any;
+
+  /** Emits errors during saving the instance. */
+  @Output() error = new EventEmitter();
+
+  /**
+   * Save the model instance. This should do any validation necessary and
+   * create or instance the model depending on whether the instance already existed.
+   */
+  save() {
+    let self = this;
+
+    this.def.save(this.data)
+      .catch(err => {
+        self.error.emit(err);
+      });
+  }
 }