+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './components';
 export * from './definitions';
 export * from './helpers';
-export * from './modules';
+export * from './module';