+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.2
+
+* Fix missing module export
+* Fix missing error handles for other components
+
 # 0.1.1
 
 * Fix components not being exported correctly